+++ commands/set.js
@@ -9,7 +9,6 @@ module.exports = {
   aliases: ["set", "modify"],
   usage: `\`${ops.prefix}set <setting> <value>\``,
 	guildOnly:true,
-	permissions: "MANAGE_GUILD",
 	execute(message, args) {
 		let config = {};
 		delete require.cache[require.resolve("../server/config.json")];