+++ commands/show-settings.js
@@ -7,7 +7,6 @@ module.exports = {
   aliases: ["settings", "options", "opts", "show"],
   usage: `${ops.prefix}settings`,
 	guildOnly:true,
-	permissions: "MANAGE_GUILD",
 	execute(message) {
 		let config = {};
 		delete require.cache[require.resolve("../server/config.json")];