+++ commands/stats.js
@@ -8,7 +8,6 @@ module.exports = {
   aliases: ["stats", "stat", "analytics", "analytic", "anals"],
   usage: `\`${ops.prefix}stats [statistic]\``,
 	guildOnly:true,
-	permissions: "MANAGE_GUILD",
 	execute(message, args) {
 		return new Promise((resolve) => {
 			if (args[0]){