+++ commands/ver.js
@@ -6,6 +6,7 @@ module.exports = {
 	description: `Tells you the current version of the bot. (If I have remembered to update it). Current version is \`${ver}\``,
   aliases: ["ver"],
   usage: `\`${ops.prefix}ver\``,
+	permissions: "VIEW_CHANNEL",
 	execute(message) {
     return new Promise(function(resolve) {
 			replyNoMention(message, `Current version: \`${ver}\``);